User can create one review per book

+++ BookHub.Server/BookHub.Server/Features/Review/Web/Models/CreateReviewWebModel.cs
@@ -11,7 +11,7 @@
         public string Content { get; set; } = null!;
 
         [Range(RatingMinValue, RatingMaxValue)]
-        public double Rating { get; init; }
+        public int Rating { get; init; }
 
         public int BookId { get; init; }
     }
